+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 * feat(scheduler): add cron with signal handling (#1) (@trants)
 * feat(cli): add version command and documentation (#4) (@trants)
 
+### Enhancements
+
+* refactor(logging): implement structured logging (#4) (@trants)
+
 ### Bug fixes
 
 * fix(ci): handle universal binary naming (#3) (@trants)